added PMD rule ClassNamingConventions
Utility classes are excluded at the moment
