+++ src/main/java/pl/betoncraft/betonquest/compatibility/jobsreborn/ConditionHasJob.java
@@ -23,17 +23,16 @@ import com.gamingmesh.jobs.container.Job;
 import com.gamingmesh.jobs.container.JobProgression;
 import org.bukkit.entity.Player;
 import pl.betoncraft.betonquest.Instruction;
-import pl.betoncraft.betonquest.api.QuestEvent;
+import pl.betoncraft.betonquest.api.Condition;
 import pl.betoncraft.betonquest.exceptions.InstructionParseException;
 import pl.betoncraft.betonquest.utils.PlayerConverter;
 
 import java.util.List;
 
-public class Event_AddExp extends QuestEvent {
+public class ConditionHasJob extends Condition {
     private String sJobName;
-    private double nAddExperience;
 
-    public Event_AddExp(final Instruction instruction) throws InstructionParseException {
+    public ConditionHasJob(final Instruction instruction) throws InstructionParseException {
         super(instruction, true);
         if (instruction.size() < 2) {
             throw new InstructionParseException("Not enough arguments");
@@ -41,11 +40,6 @@ public class Event_AddExp extends QuestEvent {
         for (final Job job : Jobs.getJobs()) {
             if (job.getName().equalsIgnoreCase(instruction.getPart(1))) {
                 sJobName = job.getName();
-                try {
-                    this.nAddExperience = Double.parseDouble(instruction.getPart(2));
-                } catch (Exception err) {
-                    throw new InstructionParseException("NUJobs_AddExp: Unable to parse the experience amount");
-                }
                 return;
             }
         }
@@ -53,16 +47,16 @@ public class Event_AddExp extends QuestEvent {
     }
 
     @Override
-    protected Void execute(final String playerID) {
+    protected Boolean execute(final String playerID) {
         final Player oPlayer = PlayerConverter.getPlayer(playerID);
 
         final List<JobProgression> oJobs = Jobs.getPlayerManager().getJobsPlayer(oPlayer).getJobProgression();
         for (final JobProgression oJob : oJobs) {
             if (oJob.getJob().getName().equalsIgnoreCase(sJobName)) {
                 //User has the job, return true
-                oJob.addExperience(nAddExperience);
+                return true;
             }
         }
-        return null;
+        return false;
     }
 }
