+++ src/main/java/pl/betoncraft/betonquest/compatibility/jobsreborn/EventAddExp.java
@@ -20,37 +20,47 @@ package pl.betoncraft.betonquest.compatibility.jobsreborn;
 
 import com.gamingmesh.jobs.Jobs;
 import com.gamingmesh.jobs.container.Job;
+import com.gamingmesh.jobs.container.JobProgression;
 import org.bukkit.entity.Player;
 import pl.betoncraft.betonquest.Instruction;
 import pl.betoncraft.betonquest.api.QuestEvent;
 import pl.betoncraft.betonquest.exceptions.InstructionParseException;
 import pl.betoncraft.betonquest.utils.PlayerConverter;
 
-public class Event_JoinJob extends QuestEvent {
+import java.util.List;
+
+public class EventAddExp extends QuestEvent {
     private String sJobName;
+    private double nAddExperience;
 
-    public Event_JoinJob(final Instruction instructions) throws InstructionParseException {
-        super(instructions, true);
-
-        if (instructions.size() < 2) {
+    public EventAddExp(final Instruction instruction) throws InstructionParseException {
+        super(instruction, true);
+        if (instruction.size() < 2) {
             throw new InstructionParseException("Not enough arguments");
         }
         for (final Job job : Jobs.getJobs()) {
-            if (job.getName().equalsIgnoreCase(instructions.getPart(1))) {
+            if (job.getName().equalsIgnoreCase(instruction.getPart(1))) {
                 sJobName = job.getName();
+                try {
+                    this.nAddExperience = Double.parseDouble(instruction.getPart(2));
+                } catch (Exception err) {
+                    throw new InstructionParseException("NUJobs_AddExp: Unable to parse the experience amount");
+                }
                 return;
             }
         }
-        throw new InstructionParseException("Jobs Reborn job " + instructions.getPart(1) + " does not exist");
+        throw new InstructionParseException("Jobs Reborn job " + instruction.getPart(1) + " does not exist");
     }
 
     @Override
     protected Void execute(final String playerID) {
         final Player oPlayer = PlayerConverter.getPlayer(playerID);
-        for (final Job job : Jobs.getJobs()) {
-            if (job.getName().equalsIgnoreCase(sJobName)) {
-                Jobs.getPlayerManager().getJobsPlayer(oPlayer).joinJob(job);
-                return null;
+
+        final List<JobProgression> oJobs = Jobs.getPlayerManager().getJobsPlayer(oPlayer).getJobProgression();
+        for (final JobProgression oJob : oJobs) {
+            if (oJob.getJob().getName().equalsIgnoreCase(sJobName)) {
+                //User has the job, return true
+                oJob.addExperience(nAddExperience);
             }
         }
         return null;
