+++ src/main/java/pl/betoncraft/betonquest/compatibility/jobsreborn/EventAddLevel.java
@@ -20,21 +20,32 @@ package pl.betoncraft.betonquest.compatibility.jobsreborn;
 
 import com.gamingmesh.jobs.Jobs;
 import com.gamingmesh.jobs.container.Job;
+import com.gamingmesh.jobs.container.JobProgression;
+import org.bukkit.entity.Player;
 import pl.betoncraft.betonquest.Instruction;
-import pl.betoncraft.betonquest.api.Condition;
+import pl.betoncraft.betonquest.api.QuestEvent;
 import pl.betoncraft.betonquest.exceptions.InstructionParseException;
+import pl.betoncraft.betonquest.utils.PlayerConverter;
 
-public class Condition_JobFull extends Condition {
+import java.util.List;
+
+public class EventAddLevel extends QuestEvent {
     private String sJobName;
+    private Integer nAddLevel;
 
-    public Condition_JobFull(final Instruction instruction) throws InstructionParseException {
+    public EventAddLevel(final Instruction instruction) throws InstructionParseException {
         super(instruction, true);
-        if (instruction.size() < 2) {
+        if (instruction.size() < 3) {
             throw new InstructionParseException("Not enough arguments");
         }
         for (final Job job : Jobs.getJobs()) {
             if (job.getName().equalsIgnoreCase(instruction.getPart(1))) {
                 sJobName = job.getName();
+                try {
+                    this.nAddLevel = Integer.parseInt(instruction.getPart(2));
+                } catch (Exception e) {
+                    throw new InstructionParseException("NUJobs_AddLevel: Unable to parse the level amount", e);
+                }
                 return;
             }
         }
@@ -42,15 +53,18 @@ public class Condition_JobFull extends Condition {
     }
 
     @Override
-    protected Boolean execute(final String playerID) {
-        for (final Job job : Jobs.getJobs()) {
-            if (job.getName().equalsIgnoreCase(sJobName)) {
-                if (job.getMaxSlots() == null)
-                    return false;
-                if (job.getTotalPlayers() >= job.getMaxSlots())
-                    return true;
+    protected Void execute(final String playerID) {
+        final Player oPlayer = PlayerConverter.getPlayer(playerID);
+
+        final List<JobProgression> oJobs = Jobs.getPlayerManager().getJobsPlayer(oPlayer).getJobProgression();
+        for (final JobProgression oJob : oJobs) {
+            if (oJob.getJob().getName().equalsIgnoreCase(sJobName)) {
+                //User has the job, return true
+                oJob.setLevel(this.nAddLevel + oJob.getLevel());
+                if (oJob.getLevel() > oJob.getJob().getMaxLevel())
+                    oJob.getJob().getMaxLevel();
             }
         }
-        return false;
+        return null;
     }
 }
