+++ src/main/java/pl/betoncraft/betonquest/compatibility/jobsreborn/EventJoinJob.java
@@ -20,52 +20,39 @@ package pl.betoncraft.betonquest.compatibility.jobsreborn;
 
 import com.gamingmesh.jobs.Jobs;
 import com.gamingmesh.jobs.container.Job;
-import com.gamingmesh.jobs.container.JobProgression;
 import org.bukkit.entity.Player;
 import pl.betoncraft.betonquest.Instruction;
-import pl.betoncraft.betonquest.api.Condition;
+import pl.betoncraft.betonquest.api.QuestEvent;
 import pl.betoncraft.betonquest.exceptions.InstructionParseException;
 import pl.betoncraft.betonquest.utils.PlayerConverter;
 
-import java.util.List;
-
-public class Condition_JobLevel extends Condition {
+public class EventJoinJob extends QuestEvent {
     private String sJobName;
-    private int nMinLevel;
-    private int nMaxLevel;
 
-    public Condition_JobLevel(final Instruction instruction) throws InstructionParseException {
-        super(instruction, true);
-        if (instruction.size() < 2) {
+    public EventJoinJob(final Instruction instructions) throws InstructionParseException {
+        super(instructions, true);
+
+        if (instructions.size() < 2) {
             throw new InstructionParseException("Not enough arguments");
         }
         for (final Job job : Jobs.getJobs()) {
-            if (job.getName().equalsIgnoreCase(instruction.getPart(1))) {
+            if (job.getName().equalsIgnoreCase(instructions.getPart(1))) {
                 sJobName = job.getName();
-                try {
-                    this.nMinLevel = Integer.parseInt(instruction.getPart(2));
-                    this.nMaxLevel = Integer.parseInt(instruction.getPart(3));
-                } catch (Exception e) {
-                    throw new InstructionParseException("NUJobs_Joblevel: Unable to parse the min or max level", e);
-                }
                 return;
             }
         }
-        throw new InstructionParseException("Jobs Reborn job " + instruction.getPart(1) + " does not exist");
+        throw new InstructionParseException("Jobs Reborn job " + instructions.getPart(1) + " does not exist");
     }
 
     @Override
-    protected Boolean execute(final String playerID) {
+    protected Void execute(final String playerID) {
         final Player oPlayer = PlayerConverter.getPlayer(playerID);
-
-        final List<JobProgression> oJobs = Jobs.getPlayerManager().getJobsPlayer(oPlayer).getJobProgression();
-        for (final JobProgression oJob : oJobs) {
-            if (oJob.getJob().getName().equalsIgnoreCase(sJobName)) {
-                //User has the job, return true
-                if (oJob.getLevel() >= nMinLevel && oJob.getLevel() <= nMaxLevel)
-                    return true;
+        for (final Job job : Jobs.getJobs()) {
+            if (job.getName().equalsIgnoreCase(sJobName)) {
+                Jobs.getPlayerManager().getJobsPlayer(oPlayer).joinJob(job);
+                return null;
             }
         }
-        return false;
+        return null;
     }
 }
