+++ src/main/java/pl/betoncraft/betonquest/compatibility/jobsreborn/JobsRebornIntegrator.java
@@ -36,26 +36,26 @@ public class JobsRebornIntegrator implements Integrator {
     @Override
     public void hook() {
         //Register conditions with beton
-        plugin.registerConditions("nujobs_canlevel", Condition_CanLevel.class);
-        plugin.registerConditions("nujobs_hasjob", Condition_HasJob.class);
-        plugin.registerConditions("nujobs_jobfull", Condition_JobFull.class);
-        plugin.registerConditions("nujobs_joblevel", Condition_JobLevel.class);
+        plugin.registerConditions("nujobs_canlevel", ConditionCanLevel.class);
+        plugin.registerConditions("nujobs_hasjob", ConditionHasJob.class);
+        plugin.registerConditions("nujobs_jobfull", ConditionJobFull.class);
+        plugin.registerConditions("nujobs_joblevel", ConditionJobLevel.class);
         LogUtils.getLogger().log(Level.INFO, "Registered Conditions [nujobs_canlevel,nujobs_hasjob,nujobs_jobfull,nujobs_joblevel]");
 
         //register events
-        plugin.registerEvents("nujobs_addexp", Event_AddExp.class);
-        plugin.registerEvents("nujobs_addlevel", Event_AddLevel.class);
-        plugin.registerEvents("nujobs_dellevel", Event_DelLevel.class);
-        plugin.registerEvents("nujobs_joinjob", Event_JoinJob.class);
-        plugin.registerEvents("nujobs_leavejob", Event_LeaveJob.class);
-        plugin.registerEvents("nujobs_setlevel", Event_SetLevel.class);
+        plugin.registerEvents("nujobs_addexp", EventAddExp.class);
+        plugin.registerEvents("nujobs_addlevel", EventAddLevel.class);
+        plugin.registerEvents("nujobs_dellevel", EventDelLevel.class);
+        plugin.registerEvents("nujobs_joinjob", EventJoinJob.class);
+        plugin.registerEvents("nujobs_leavejob", EventLeaveJob.class);
+        plugin.registerEvents("nujobs_setlevel", EventSetLevel.class);
         LogUtils.getLogger().log(Level.INFO, "Registered Events [nujobs_addexp,nujobs_addlevel,nujobs_dellevel,nujobs_joinjob,nujobs_leavejob,nujobs_setlevel]");
 
         //register objectives
-        plugin.registerObjectives("nujobs_joinjob", Objective_JoinJob.class);
-        plugin.registerObjectives("nujobs_leavejob", Objective_LeaveJob.class);
-        plugin.registerObjectives("nujobs_levelup", Objective_LevelUpEvent.class);
-        plugin.registerObjectives("nujobs_payment", Objective_PaymentEvent.class);
+        plugin.registerObjectives("nujobs_joinjob", ObjectiveJoinJob.class);
+        plugin.registerObjectives("nujobs_leavejob", ObjectiveLeaveJob.class);
+        plugin.registerObjectives("nujobs_levelup", ObjectiveLevelUpEvent.class);
+        plugin.registerObjectives("nujobs_payment", ObjectivePaymentEvent.class);
         LogUtils.getLogger().log(Level.INFO, "Registered Objectives [nujobs_joinjob,nujobs_leavejob,nujobs_levelup,nujobs_payment]");
 
     }
